requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.2
Flask==2.2.5
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mutagen==1.46.0
pycryptodomex==3.18.0
pysondb-v2==2.0.0
redis==4.5.5
requests==2.31.0
requests-file==1.5.1
selenium==3.14.1
selenium-requests==1.3
six==1.16.0
soupsieve==2.4.1
tldextract==3.4.4
typing_extensions==4.6.3
urllib3==1.26.2
websockets==11.0.3
Werkzeug==2.2.3
wget==3.2
yt-dlp==2023.3.4
zipp==3.15.0
